requirements_nibi.txt
accelerate==1.8.1+computecanada
aiofiles==23.2.1+computecanada
aiohappyeyeballs==2.6.1+computecanada
aiohttp==3.12.13+computecanada
aiosignal==1.4.0+computecanada
annotated_types==0.7.0+computecanada
anyio==4.9.0+computecanada
asttokens==3.0.0+computecanada
attrs==25.3.0+computecanada
auto_gptq==0.7.1+computecanada
autopep8==2.3.2
boto3==1.40.2
botocore==1.40.2
Brotli==1.1.0+computecanada
certifi==2025.6.15+computecanada
charset_normalizer==3.4.2+computecanada
click==8.2.1+computecanada
comm==0.2.2+computecanada
contourpy==1.3.1+computecanada
cycler==0.12.1+computecanada
datasets==4.0.0
debugpy==1.8.12+computecanada
decorator==5.1.1+computecanada
device-smi==0.4.1
dill==0.3.8+computecanada
distro==1.9.0+computecanada
dotenv==0.9.9
executing==2.2.0+computecanada
fastapi==0.115.12+computecanada
ffmpy==0.5.0+computecanada
filelock==3.18.0+computecanada
fonttools==4.55.8+computecanada
frozenlist==1.7.0+computecanada
fsspec==2025.3.0+computecanada
gekko==1.3.0
gptqmodel==2.2.0
gradio==5.39.0
gradio_client==1.11.0
groovy==0.1.2+computecanada
h11==0.16.0+computecanada
hf_transfer==0.1.9+computecanada
hf_xet==1.1.3+computecanada
httpcore==1.0.9+computecanada
httpx==0.28.1+computecanada
huggingface-hub==0.34.3
idna==3.10+computecanada
iniconfig==2.1.0+computecanada
ipykernel==6.29.5+computecanada
ipython==8.32.0+computecanada
ipywidgets==8.1.5+computecanada
jedi==0.19.2+computecanada
jinja2==3.1.6+computecanada
jiter==0.6.1+computecanada
jmespath==1.0.1+computecanada
joblib==1.5.1+computecanada
jupyter_client==8.6.3+computecanada
jupyter_core==5.7.2+computecanada
jupyterlab_widgets==3.0.13+computecanada
kiwisolver==1.4.8+computecanada
logbar==0.0.4
markdown_it_py==3.0.0+computecanada
MarkupSafe==2.1.5+computecanada
matplotlib==3.10.0+computecanada
matplotlib_inline==0.1.7+computecanada
mdurl==0.1.2+computecanada
mpmath==1.3.0+computecanada
multidict==6.6.3+computecanada
multiprocess==0.70.16+computecanada
nest_asyncio==1.6.0+computecanada
networkx==3.5+computecanada
nltk==3.9.1+computecanada
nose==1.3.7+computecanada
numpy==1.26.4+computecanada
openai==1.90.0+computecanada
optimum==1.26.1
orjson==3.10.5+computecanada
packaging==25.0+computecanada
pandas==2.2.3+computecanada
parso==0.8.4+computecanada
peft==0.16.0
pexpect==4.9.0+computecanada
pillow==11.1.0+computecanada
platformdirs==3.10.0+computecanada
pluggy==1.6.0+computecanada
prompt_toolkit==3.0.50+computecanada
propcache==0.3.2+computecanada
protobuf==6.31.1+computecanada
psutil==6.1.1+computecanada
ptyprocess==0.7.0+computecanada
pure_eval==0.2.3+computecanada
pyarrow @ file:///tmp/ebuser/avx512/Arrow/19.0.1/GCCcore-12.3-gentoo/arrow/python
pycodestyle==2.14.0
pydantic==2.11.7+computecanada
pydantic_core==2.33.2+computecanada
pydub==0.25.1+computecanada
pygments==2.19.2+computecanada
pyparsing==3.2.1+computecanada
pytest==8.4.1+computecanada
python_dateutil==2.9.0.post0+computecanada
python_dotenv==1.1.1+computecanada
python_multipart==0.0.20+computecanada
pytorch-pretrained-biggan==0.1.1
pytz==2025.2+computecanada
PyYAML==6.0.2+computecanada
pyzmq==26.2.1+computecanada
random_word==1.0.13
regex==2024.11.6+computecanada
requests==2.32.4+computecanada
rich==14.0.0+computecanada
rouge==1.0.1+computecanada
ruff==0.11.0+computecanada
s3transfer==0.13.1
safehttpx==0.1.6+computecanada
safetensors==0.5.3+computecanada
scikit_learn==1.6.1+computecanada
scipy==1.15.1+computecanada
seaborn==0.13.2+computecanada
semantic_version==2.10.0+computecanada
sentencepiece==0.2.0+computecanada
shellingham==1.5.4+computecanada
six==1.17.0+computecanada
sklearn==0.0+computecanada
sniffio==1.3.1+computecanada
stack_data==0.6.3+computecanada
starlette==0.46.2+computecanada
sympy==1.14.0+computecanada
threadpoolctl==3.6.0+computecanada
tokenicer==0.0.4
tokenizers==0.21.1+computecanada
tomlkit==0.13.3+computecanada
torch==2.7.1+computecanada
tornado==6.4.2+computecanada
tqdm==4.67.1+computecanada
traitlets==5.14.3+computecanada
transformers==4.53.3
typer==0.16.0+computecanada
typing_extensions==4.14.0+computecanada
typing_inspection==0.4.1+computecanada
tzdata==2025.2+computecanada
urllib3==2.5.0+computecanada
uvicorn==0.34.3+computecanada
wcwidth==0.2.13+computecanada
websockets==15.0.1+computecanada
widgetsnbextension==4.0.13+computecanada
xxhash==3.5.0+computecanada
yarl==1.20.1+computecanada
